requirements.txt
# Aggregated Python dependencies for the project
# Sourced from backend/requirements.txt and project usage (channels, channels-redis, redis)

asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
gunicorn
pandas
whitenoise
python-dotenv==1.2.1

# WebSocket / Channels
channels
channels-redis
redis

# Optional / dev (uncomment to install)
# pytest
# pytest-django
# coverage
